refactor camera live mode hook
